byte값을 소숫점 2번째짜리까지 끊어서 보여주도록 변경
