+++ portal/Portal/sources/views/dash/network.js
@@ -10,6 +10,7 @@
 */
 import { JetView } from "webix-jet";
 import { sizeToString } from "../../models/utils/sizeToString";
+import { sizeToHtml} from "../../models/utils/sizeToHtml";
 
 const MY_RX_TITLE = "io_tx_title";
 const MY_TX_TITLE = "io_rx_title";
@@ -151,40 +152,12 @@ function loadNetWorkUsages() {
 					});
 					$$(MY_RX).add(usage);
 					$$(MY_TX).add(usage);
-					$$(MY_RX_TITLE).setValue(sizeToHtml(usage.Rx, "Rx"));
-					$$(MY_TX_TITLE).setValue(sizeToHtml(usage.Tx, "Tx"));
+					$$(MY_RX_TITLE).setValue(sizeToHtml(usage.Rx, "Read"));
+					$$(MY_TX_TITLE).setValue(sizeToHtml(usage.Tx, "Write"));
 				}
 			},
 			function (error) {
 				stopTimer();
 			}
 		);
-}
-
-/**
- * 입력된 byte를 보기쉬운 용량 단위로 변환하여 반환한다.
- * @param {long} bytes 변환할 용량
- * @param {String} type 타입
- * @returns 용량 + 단위
- */
-export function sizeToHtml(bytes, type) {
-	const thresh = 1000;
-	const dp = 1;
-
-	if (Math.abs(bytes) < thresh) {
-		if (type == "Rx") return `<span class='io_in'>${bytes}</span><span class='io_in_unit'>B/s</span>`;
-		else return `<span class='io_out'>${bytes}</span><span class='io_out_unit'>B/s</span>`;
-	}
-
-	const units = ["kB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"];
-	let u = -1;
-	const r = 10 ** dp;
-
-	do {
-		bytes /= thresh;
-		++u;
-	} while (Math.round(Math.abs(bytes) * r) / r >= thresh && u < units.length - 1);
-
-	if (type == "Rx") return `<span class='io_in'>${bytes.toFixed(dp)}</span><span class='io_in_unit'>${units[u]}/s</span>`;
-	else return `<span class='io_out'>${bytes.toFixed(dp)}</span><span class='io_out_unit'>${units[u]}/s</span>`;
 }
